backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.82.0
qdrant-client==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pydantic-settings==2.7.1
markdown==3.7.0
tiktoken==0.9.0
sse-starlette==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
